% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distbuilder-package.R
\docType{package}
\name{distbuilder-package}
\alias{distbuilder}
\alias{distbuilder-package}
\title{distbuilder: Build samplers from distribution specifications}
\description{
A small package for constructing reusable random sampler functions from
distribution specifications defined as R lists or YAML.
}
\details{
All samplers are created using \code{build_sampler()}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CamNZ/distbuilder}
  \item Report bugs at \url{https://github.com/CamNZ/distbuilder/issues}
}

}
\author{
\strong{Maintainer}: Cameron Fraser \email{noreply@github.com}

Authors:
\itemize{
  \item Cameron Fraser \email{noreply@github.com}
}

}
\keyword{internal}
